requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
langchain==0.3.25
langchain-google-genai==2.1.5
python-dotenv==1.0.0
